+++ composeApp/src/commonMain/kotlin/ua/vald_zx/game/rat/race/card/beans/Models.kt
@@ -5,12 +5,12 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ProfessionCard(
     val profession: String = "",
-    val salary: Int = 0,
-    val rent: Int = 0,
-    val food: Int = 0,
-    val cloth: Int = 0,
-    val transport: Int = 0,
-    val phone: Int = 0,
+    val salary: Long = 0,
+    val rent: Long = 0,
+    val food: Long = 0,
+    val cloth: Long = 0,
+    val transport: Long = 0,
+    val phone: Long = 0,
 )
 
 enum class BusinessType {
@@ -24,9 +24,9 @@ enum class BusinessType {
 data class Business(
     val type: BusinessType,
     val name: String,
-    val price: Int,
-    val profit: Int,
-    val extentions: List<Int> = emptyList()
+    val price: Long,
+    val profit: Long,
+    val extentions: List<Long> = emptyList()
 )
 
 enum class SharesType {
@@ -39,9 +39,9 @@ enum class SharesType {
 @Serializable
 data class Shares(
     val type: SharesType,
-    val count: Int,
-    val buyPrice: Int,
+    val count: Long,
+    val buyPrice: Long,
 ) {
-    val price: Int
+    val price: Long
         get() = count * buyPrice
 }
